move browser drawer stuff into subdir
